+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Fixed
+- https://github.com/StateSmith/StateSmith/issues/49
+
+### Changed
+- Transition actions are now run after states are exited (instead of before) https://github.com/StateSmith/StateSmith/issues/6
+
+### Added
+- Support initial/entry/exit multiple https://github.com/StateSmith/StateSmith/issues/40
+
+
+
+
 ## [0.5.9-alpha]
 ### Added
 - Runner - set process exit code to -1 on failure.